requirements.txt
# pyaudio
# vosk
# channels[daphne]
# kokoro>=0.9.2 
# soundfile
# ffmpeg
# grpcio 
# grpcio-tools
# protobuf
# huggingface_hub[hf_xet]

Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-extensions
drf-spectacular
pillow
drf-api-logger
django-schema-viewer
Faker
python-dotenv
loguru
daphne
channels 

## srinidhi
uvicorn
fastapi
psycopg[binary]
psycopg2
flask_sqlalchemy